+++ src/controllers/visualization.controller.ts
@@ -10,6 +10,7 @@ import { Status } from "../models/status.enum";
 import { getMssqlDbObjects, getMssqlDbColumns, getMssqlDbColumnValues, getMssqlDbTableRows } from "../repositories/mssql.repository";
 import { ViewModel } from "../models/view.model";
 import { ConnectionContext } from '../connection-context';
+import { FQName } from '../FQName';
 
 interface IIncomingMessage {
     command: string;
@@ -59,9 +60,10 @@ const renderWebviewContent = async (webview: vscode.Webview, connectionContext: 
     if (connectionContext.connection?.options.database && connectionContext.connectionId) {
 
         let viewModel = new ViewModel();
-        viewModel.serverName = connectionContext.fqname?.serverName;
-        viewModel.databaseName = connectionContext.fqname?.databaseName;
-        viewModel.filterObjectsSchema = connectionContext.fqname?.schemaName;
+        viewModel.serverName = connectionContext.fqname.serverName;
+        viewModel.databaseName = connectionContext.fqname.databaseName;
+        viewModel.filterObjectsSchema = connectionContext.fqname.schemaName;
+        viewModel.selectedObjectName = connectionContext.fqname.tableName;
         viewModel.autoApply = true;
         viewModel.selectTables = true;
         viewModel.selectViews = true;
@@ -84,6 +86,11 @@ const renderWebviewContent = async (webview: vscode.Webview, connectionContext: 
                     }
                     case 'loadObjects': {
                         await loadObjects(connectionId, webview, viewModel);
+                        // also load the columns and rows when starting with a selected table
+                        if (connectionContext.fqname.tableName !== undefined) {
+                            await loadColumns(connectionId, webview, viewModel);
+                            await loadRows(connectionId, webview, viewModel);
+                        }
                         break;
                     }
                     case 'loadColumns': {
@@ -244,7 +251,6 @@ const loadObjects = async (connectionId: string, webview: azdata.DashboardWebvie
     viewModel.rowsColumnsName = undefined;
     viewModel.rows = undefined;
     viewModel.rowsCount = undefined;
-    //TODO index
     viewModel.selectedObjectIndex = undefined;
     viewModel.selectedColumnIndex = undefined;
     viewModel.selectedValueIndex = undefined;
@@ -255,6 +261,10 @@ const loadObjects = async (connectionId: string, webview: azdata.DashboardWebvie
     if (viewModel.filterObjectsSchema !== undefined
         && viewModel.filterObjectsSchema !== '*') {
         viewModel.objects = viewModel.objects.filter(t => t.Schema === viewModel.filterObjectsSchema);
+        if (viewModel.selectedObjectName !== undefined) {
+            let idx = viewModel.objects.findIndex(o => o.Schema === viewModel.filterObjectsSchema && o.Name === viewModel.selectedObjectName);
+            viewModel.selectedObjectIndex = idx >= 0 ? idx : undefined;
+        }
 
         //viewModel.columns = undefined;
         //viewModel.values = undefined;
@@ -272,6 +282,7 @@ const loadObjects = async (connectionId: string, webview: azdata.DashboardWebvie
             objects: viewModel.objects,
             objectsSchema: viewModel.objectsSchema,
             filterObjectsSchema: viewModel.filterObjectsSchema,
+            objectIndex: viewModel.selectedObjectIndex,
             columns: undefined,
             values: undefined,
             rows: undefined